🧪 WRTQP0 hooks: split pre-push binding tests
