+++ packages/agentplane/src/cli/run-cli.core.hooks.hook-run.test.ts
@@ -442,252 +442,6 @@ describe("runCli hooks run", { timeout: HOOKS_SUITE_TIMEOUT_MS }, () => {
     }
   });
 
-  it("hooks run pre-push dispatches the real script", async () => {
-    const root = await mkGitRepoRoot();
-    await writeDefaultConfig(root);
-    await writeFile(
-      path.join(root, "package.json"),
-      JSON.stringify(
-        {
-          name: "hook-test",
-          private: true,
-          scripts: {
-            "format:check": "node scripts/format-check.mjs",
-            "ci:local:fast": "node scripts/ci-fast.mjs",
-            "ci:local:full": "node scripts/ci-fast.mjs",
-          },
-        },
-        null,
-        2,
-      ),
-      "utf8",
-    );
-    await mkdir(path.join(root, "scripts"), { recursive: true });
-    await mkdir(path.join(root, "src"), { recursive: true });
-    await writeFile(path.join(root, "scripts", "format-check.mjs"), "process.exit(1);\n", "utf8");
-    await writeFile(path.join(root, "scripts", "ci-fast.mjs"), "process.exit(0);\n", "utf8");
-    await writeFile(path.join(root, "src", "example.ts"), "export const example = 1;\n", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "package.json", "scripts", "src/example.ts"], { cwd: root });
-    await execFileAsync("git", ["commit", "-m", "test fixture"], { cwd: root });
-
-    const io = captureStdIO();
-    try {
-      const code = await runCli(["hooks", "run", "pre-push", "--root", root]);
-      expect(code).toBe(1);
-    } finally {
-      io.restore();
-    }
-  });
-
-  it("pre-push release mode reports polluted local git config before payload checks", async () => {
-    const root = await mkGitRepoRoot();
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["config", "--local", "core.bare", "true"], {
-      cwd: root,
-      env: cleanGitEnv(),
-    });
-
-    let failure: (Error & { stderr?: string | Buffer; stdout?: string | Buffer }) | null = null;
-    try {
-      execFileSync("node", [PRE_PUSH_HOOK_SCRIPT], {
-        cwd: root,
-        stdio: "pipe",
-        input: "",
-        env: {
-          ...process.env,
-          AGENTPLANE_HOOKS_RELEASE: "1",
-        },
-      });
-    } catch (error) {
-      failure = error as Error & { stderr?: string | Buffer; stdout?: string | Buffer };
-    }
-
-    expect(failure).not.toBeNull();
-    expect(String(failure?.stdout ?? "")).toContain("Running pre-push checks in release mode.");
-    expect(String(failure?.stderr ?? "")).toContain(
-      "pre-push blocked: release checks cannot run because local git config has core.bare=true.",
-    );
-    expect(String(failure?.stderr ?? "")).toContain(
-      "This indicates repository config pollution, not a release payload failure.",
-    );
-    expect(String(failure?.stdout ?? "")).not.toContain("== Format (check) ==");
-  });
-
-  it("pre-push hook blocks outgoing mutating commits without task binding", async () => {
-    const root = await mkGitRepoRootWithBranch("main");
-    await configureGitUser(root);
-    await writeDefaultConfig(root);
-    await writeFile(
-      path.join(root, "package.json"),
-      JSON.stringify(
-        {
-          name: "hook-test",
-          private: true,
-          scripts: {
-            "format:check": 'node -e "process.exit(0)"',
-            "ci:local:fast": 'node -e "process.exit(0)"',
-          },
-        },
-        null,
-        2,
-      ),
-      "utf8",
-    );
-    await commitAll(root, "chore: base");
-    const baseSha = execFileSync("git", ["rev-parse", "HEAD"], {
-      cwd: root,
-      encoding: "utf8",
-    }).trim();
-
-    await mkdir(path.join(root, "src"), { recursive: true });
-    await writeFile(path.join(root, "src", "app.ts"), "export const value = 1;\n", "utf8");
-    await commitAll(root, "✨ code: connect managed API");
-    const headSha = execFileSync("git", ["rev-parse", "HEAD"], {
-      cwd: root,
-      encoding: "utf8",
-    }).trim();
-
-    let failure: (Error & { stderr?: string | Buffer; stdout?: string | Buffer }) | null = null;
-    try {
-      execFileSync("node", [PRE_PUSH_HOOK_SCRIPT], {
-        cwd: root,
-        stdio: "pipe",
-        input: `refs/heads/main ${headSha} refs/heads/main ${baseSha}\n`,
-      });
-    } catch (error) {
-      failure = error as Error & { stderr?: string | Buffer; stdout?: string | Buffer };
-    }
-
-    expect(failure).not.toBeNull();
-    expect(String(failure?.stderr ?? "")).toContain(
-      "pre-push blocked: mutating commits require a valid task id",
-    );
-    expect(String(failure?.stderr ?? "")).toContain("src/app.ts");
-  });
-
-  it("pre-push hook includes merge commits in mutating commit audits", async () => {
-    const root = await mkGitRepoRootWithBranch("main");
-    await configureGitUser(root);
-    await writeDefaultConfig(root);
-    await writeFile(
-      path.join(root, "package.json"),
-      JSON.stringify(
-        {
-          name: "hook-test",
-          private: true,
-          scripts: {
-            "format:check": 'node -e "process.exit(0)"',
-            "ci:local:fast": 'node -e "process.exit(0)"',
-          },
-        },
-        null,
-        2,
-      ),
-      "utf8",
-    );
-    await mkdir(path.join(root, ".agentplane", "tasks", "202601010101-ABCDEF"), {
-      recursive: true,
-    });
-    await writeFile(
-      path.join(root, ".agentplane", "tasks", "202601010101-ABCDEF", "README.md"),
-      "task\n",
-      "utf8",
-    );
-    await commitAll(root, "chore: base");
-    const baseSha = execFileSync("git", ["rev-parse", "HEAD"], {
-      cwd: root,
-      encoding: "utf8",
-    }).trim();
-
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["checkout", "-b", "feature"], { cwd: root });
-    await mkdir(path.join(root, "src"), { recursive: true });
-    await writeFile(path.join(root, "src", "app.ts"), "export const value = 1;\n", "utf8");
-    await commitAll(root, "✨ ABCDEF code: connect managed API");
-    await execFileAsync("git", ["checkout", "main"], { cwd: root });
-    await execFileAsync("git", ["merge", "--no-ff", "feature", "-m", "Merge feature branch"], {
-      cwd: root,
-    });
-    const headSha = execFileSync("git", ["rev-parse", "HEAD"], {
-      cwd: root,
-      encoding: "utf8",
-    }).trim();
-
-    let failure: (Error & { stderr?: string | Buffer; stdout?: string | Buffer }) | null = null;
-    try {
-      execFileSync("node", [PRE_PUSH_HOOK_SCRIPT], {
-        cwd: root,
-        stdio: "pipe",
-        input: `refs/heads/main ${headSha} refs/heads/main ${baseSha}\n`,
-      });
-    } catch (error) {
-      failure = error as Error & { stderr?: string | Buffer; stdout?: string | Buffer };
-    }
-
-    expect(failure).not.toBeNull();
-    expect(String(failure?.stderr ?? "")).toContain("Merge feature branch");
-    expect(String(failure?.stderr ?? "")).toContain("src/app.ts");
-  });
-
-  it("pre-push hook accepts emergency hotfix commits with backfill evidence", async () => {
-    const root = await mkGitRepoRootWithBranch("main");
-    await configureGitUser(root);
-    await writeDefaultConfig(root);
-    await writeFile(
-      path.join(root, "package.json"),
-      JSON.stringify(
-        {
-          name: "hook-test",
-          private: true,
-          scripts: {
-            "format:check": 'node -e "process.exit(0)"',
-            "ci:local:fast": 'node -e "process.exit(0)"',
-          },
-        },
-        null,
-        2,
-      ),
-      "utf8",
-    );
-    await commitAll(root, "chore: base");
-    const baseSha = execFileSync("git", ["rev-parse", "HEAD"], {
-      cwd: root,
-      encoding: "utf8",
-    }).trim();
-
-    await mkdir(path.join(root, "src"), { recursive: true });
-    await writeFile(path.join(root, "src", "app.ts"), "export const value = 1;\n", "utf8");
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", "src/app.ts"], { cwd: root });
-    await execFileAsync(
-      "git",
-      [
-        "commit",
-        "-m",
-        [
-          "🚑 hotfix: restore managed API",
-          "",
-          "Emergency-Hotfix: true",
-          "Backfill-Task: 202601010101-ABCDEF",
-          "Backfill-Evidence: incident ticket and rollback note recorded",
-        ].join("\n"),
-      ],
-      { cwd: root },
-    );
-    const headSha = execFileSync("git", ["rev-parse", "HEAD"], {
-      cwd: root,
-      encoding: "utf8",
-    }).trim();
-
-    const result = execFileSync("node", [PRE_PUSH_HOOK_SCRIPT], {
-      cwd: root,
-      encoding: "utf8",
-      input: `refs/heads/main ${headSha} refs/heads/main ${baseSha}\n`,
-    });
-    expect(result).toContain("Running pre-push checks in standard mode.");
-  });
-
   it("hooks run post-merge prunes merged local task worktrees on the base branch", async () => {
     const root = await mkGitRepoRootWithBranch("main");
     await configureGitUser(root);
